Refact: 배치 클래스 네임 수정, 패키지 구조 수정
