+++ src/main/java/com/ureca/idle/batch/kidspersonalitychangehistory/config/SaveKidsPersonalityJobConfiguration.java
@@ -1,24 +1,25 @@
-package com.ureca.idle.batch.kidsPersonalityDeleteHistory;
+package com.ureca.idle.batch.kidspersonalitychangehistory.config;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
-import org.springframework.batch.core.job.builder.JobBuilder;
 import org.springframework.batch.core.launch.support.RunIdIncrementer;
 import org.springframework.batch.core.repository.JobRepository;
+import org.springframework.batch.core.job.builder.JobBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @RequiredArgsConstructor
-public class DeleteKidsJobConfiguration {
+public class SaveKidsPersonalityJobConfiguration {
+
+    private final Step saveKidsPersonalityStep;
 
     @Bean
-    public Job deleteBatchJob(Step deleteBatchStep, JobRepository jobRepository) throws Exception {
-        System.out.println("Delete Job Start");
-        return new JobBuilder("deleteKidsPersonalityBatchJob", jobRepository)
+    public Job saveKidsPersonalityJob(JobRepository jobRepository) {
+        return new JobBuilder("saveKidsPersonalityJob", jobRepository)
                 .incrementer(new RunIdIncrementer())
-                .start(deleteBatchStep)
+                .start(saveKidsPersonalityStep)
                 .build();
     }
 }
