+++ src/main/java/com/ureca/idle/batch/kidspersonalitychangehistory/dto/KidsPersonalityChangeHistoryResp.java
@@ -1,4 +1,4 @@
-package com.ureca.idle.batch.dto;
+package com.ureca.idle.batch.kidspersonalitychangehistory.dto;
 
 public record KidsPersonalityChangeHistoryResp(Long kidsId, int ei, int sn, int tf, int jp, String mbti) {
 }
