+++ src/main/java/com/ureca/idle/batch/kidspersonalitydeletehistory/KidsPersonalityDeleteHistoryDto.java
@@ -1,4 +1,4 @@
-package com.ureca.idle.batch.kidsPersonalityDeleteHistory;
+package com.ureca.idle.batch.kidspersonalitydeletehistory;
 
 import lombok.Getter;
