+++ src/main/java/com/ureca/idle/batch/kidspersonalitydeletehistory/config/DeleteKidsJobConfiguration.java
@@ -1,25 +1,24 @@
-package com.ureca.idle.batch.kidspersonalitybatch;
+package com.ureca.idle.batch.kidspersonalitydeletehistory.config;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
+import org.springframework.batch.core.job.builder.JobBuilder;
 import org.springframework.batch.core.launch.support.RunIdIncrementer;
 import org.springframework.batch.core.repository.JobRepository;
-import org.springframework.batch.core.job.builder.JobBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @RequiredArgsConstructor
-public class SaveKidsJobConfiguration {
-
-    private final Step saveKidsPersonalityStep;
+public class DeleteKidsJobConfiguration {
 
     @Bean
-    public Job saveKidsPersonalityJob(JobRepository jobRepository) {
-        return new JobBuilder("saveKidsPersonalityJob", jobRepository)
+    public Job deleteBatchJob(Step deleteBatchStep, JobRepository jobRepository) throws Exception {
+        System.out.println("Delete Job Start");
+        return new JobBuilder("deleteKidsPersonalityBatchJob", jobRepository)
                 .incrementer(new RunIdIncrementer())
-                .start(saveKidsPersonalityStep)
+                .start(deleteBatchStep)
                 .build();
     }
 }
