+++ src/main/java/com/ureca/idle/batch/kidspersonalitydeletehistory/reader/DeleteKidsItemReader.java
@@ -1,6 +1,8 @@
-package com.ureca.idle.batch.kidsPersonalityDeleteHistory;
+package com.ureca.idle.batch.kidspersonalitydeletehistory.reader;
 
 
+import com.ureca.idle.batch.kidspersonalitydeletehistory.KidsPersonalityDeleteHistoryDto;
+import com.ureca.idle.batch.kidspersonalitydeletehistory.KidsPersonalityDeleteHistoryDtoRowMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.item.database.JdbcPagingItemReader;
 import org.springframework.batch.item.database.PagingQueryProvider;
